Remove publicPath from webpack.config.renderer.prod
Webpack was adding `/dist` to the beginning of every URL referenced relatively from `styles.css` or `renderer.prod.js`, causing 404 errors on assets referenced from those bundles.

This fix removes the need for `publicPath` to be forced to the proper value in the config for `ExtractTextPlugin`, thus fixing asset references for all transforms in the renderer prod bundle.