Added window re-activation for MacOS
On MacOS, when you close the window using the "x" close button, you cannot re-open the app unless right clicking on the app icon in dock and selecting "Quit" and then start it again. This change re-creates the window when clicking on the dock icon after the window has been soft-closed.