Add comment so folk know better how to add things
A nudge to help make sure folk don't become frustrated trying to add a new page. Issue #1767 brought me to this.
As a noob, It took me a while to find/realize this. Even better would be some documentation additions or a quick tutorial on how to add a "page".